Added files, setting up other connectors
