+++ src/connectors/coingecko/coingeckoConnector.test.ts
@@ -1,0 +1,1 @@
+import CoingeckoConnector from './coingeckoConnector';
