+++ src/connectors/exchangerates/exchangeratesConnector.test.ts
@@ -1,0 +1,1 @@
+import ExchangeratesConnector from './exchangeratesConnector';
